"Amazont" API 1 (Final)

+++ database/migrations/2025_03_06_162856_create_productos_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('nombre');
             $table->text('descripcion');
-            $table->integer('precio');
+            $table->float('precio');
             $table->string('estado');
             $table->boolean('oferta')->default(false);
             $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
